@@ -1,18 +1,18 @@
-import { Switch, Route } from 'react-router-dom';
+import { Routes, Route } from 'react-router-dom';
 
 // PAGES
 import Home from './pages/Home/Home';
 import EditContact from './pages/EditContact/EditContact';
 import NewContact from './pages/NewContact/NewContact';
 
-function Routes() {
+function ApplicationRoutes() {
   return (
-    <Switch>
-      <Route exact path="/" component={Home} />
-      <Route exact path="/new" component={NewContact} />
-      <Route exact path="/edit/:id" component={EditContact} />
-    </Switch>
+    <Routes>
+      <Route exact path="/" element={<Home />} />
+      <Route exact path="/new" element={<NewContact />} />
+      <Route exact path="/edit/:id" element={<EditContact />} />
+    </Routes>
   );
 }
 
-export default Routes;
+export default ApplicationRoutes;
